requirements.txt
discord.py >= 2.5.2
flask >= 2.3.0
python-dotenv>=1.0.0